execve: reset LD_PRELOAD if was unset
The variable LD_PRELOAD can be unset by subsequent binaries (see
dracut.sh[1]). However, it is *required* to have libiamroot.so loader
first to resolve path correclty; otherwise it ends with ENOENT errors.

For now, this resets the environment LD_PRELOAD with the content of
IAMROOT_LIB (default to /usr/lib/iamroot/libiamroot.so is unset) if the
it unset; unless a better alternative is found in the future.

[1]: https://github.com/dracutdevs/dracut/commit/45404a2ad884edac2646f5f86045247f635b5673