realpath: preserve errno
realpath calls functions which may change the value of errno.

readlink is called for every path components, sets errno to EINVAL if a
path component is not a symlink, and makes realpath return an erroneous
EINVAL errno value even if the resolution is a success.

This saves errno after enterring the function to restore it before
returning, and resets it after call to readlink.