exec.sh: add ldd
ldd is used by initramfs creation tools to get the list of libraries
linked to binaries.

ldd outputs the preloaded libraries and the full path to the linked
libraries; which includes the iamroot library (LD_PRELOAD) and the path
of the chroot'ed directory (LD_LIBRARY_PATH).

This adds ldd to the variable IAMROOT_EXEC_IGNORE to handle it in the
script exec.sh to unset both variables LD_PRELOAD and LD_LIBRARY_PATH:
this does not output the libiamroot.so library the chroot directory
$IAMROOT_ROOT in the path of linked libraries