Add padding after allocated regex_t structures
TL;DR; This commit fixes a silent overflow if running gpg from within an
iamroot environment; starting from gnupg 2.2.22.

It leads to a segmentation fault if the two destructor functions are run
in the opposite order. Consider the change below for reproduction:

	--- a/Makefile
	+++ b/Makefile
	@@ -66,6 +66,7 @@ libiamroot.so: execl.o
	 libiamroot.so: execle.o
	 libiamroot.so: execlp.o
	 libiamroot.so: execv.o
	+libiamroot.so: fpath_resolutionat.o
	 libiamroot.so: execve.o
	 libiamroot.so: execvp.o
	 libiamroot.so: execvpe.o
	@@ -77,7 +78,6 @@ libiamroot.so: fchownat.o
	 libiamroot.so: fexecve.o
	 libiamroot.so: fopen.o
	 libiamroot.so: fopen64.o
	-libiamroot.so: fpath_resolutionat.o
	 libiamroot.so: freopen.o
	 libiamroot.so: freopen64.o
	 libiamroot.so: fstat.o

And the following command to cause the segmentation fault:

	$ iamroot-shell -c gpg
	$ bash iamroot-shell -c ./gpg </dev/null
	gpg: WARNING: no command supplied.  Trying to guess what you mean ...
	gpg: processing message failed: Unknown system error

	gpg: signal 11 caught ... exiting
	Segmentation fault (core dumped)

This adds extra padding after every allocated regex_t structures to let
the overflow write beyong the data but to memory that is left unused.
Thus, it prevent writing over data that would causes behaviours that are
not expected.

The long story then.

The regex[1] POSIX API does not have a function to allocate the regex_t
object; this structure is not opaque and its size varies accross the
various implementations.

It is not big deal in a static world; but it could matter in a dynamic
system because the regex_t variables are allocated by the compiler, not
by the API itself; and the symbols that consume the data at runtime may
be different than the ones that was used for compilation.

Overflows happen if the regex symbols at runtime use regex_t variables
with a size structure that is bigger than the memory that was allocated
by the compiler.

However, such a usecase never happen in the real world, at least not
until iamroot has met the version 2.2.22 of gpg.

The GnuPG project needed regular expression support on systems that do
not have such system library for an interoperability reason. Therefore,
gnupg has integrated the jimtcl regex implementation[2].

This implementation uses a regex_t footprint that is larger that the
standard linux libc libraries; the size of the data structure is 106
Bytes long while it take 64 Bytes of memory only on glibc, uclibc and
musl libraries.

Also, the API functions held originally the same name[3] as the POSIX
API until they were prefixed by jim_[4] to resolve a symbol conflict[5];
and this rename has never been backported to gnupg then.

The overflow happen in iamroot. It is caused by the symbol regexec from
gnupg that is loaded and called in place of one from the libc and which
writes beyond the data that were allocated by the compiler.

The gpg packages are shipped by distributions, and the work is broken
already. Therefore the solution to let iamroot work correctly is to add
extra padding after every allocated regex_t structures (goto TL;DR;).

[1]: https://linux.die.net/man/3/regex
[2]: https://github.com/gpg/gnupg/commit/ba247a114c75a84473c11c1484013b09fbb9bcd1
[3]: https://github.com/msteveb/jimtcl/blob/0.79/jimregexp.h#L100-L103
[4]: https://github.com/msteveb/jimtcl/commit/795854e54681f86fe66d8dcf8b09e44b9f726a62
[5]: https://github.com/msteveb/jimtcl/issues/163