Fix alpine-test by preloading the musl library
The commit c0c61e8ac5398b80485743905f2f8c4f22ae140c makes possible to
run the dynamic loader directly; it has introduced the test of invoking
the musl loader directly from a musl rootfs which is fragile somehow.

A big part of the cake remains missing here as iamroot is not packaged
yet by the distros; the library exists on the developer's system only.
Therefore, it is the library that is compiled on the host system which
is used by dynamic loader in the chroot'ed system for now thanks to the
environment LD_PRELOAD.

The world is not ideal and the iamroot library depends on libc's symbols
on which it was built. As a consequence, doing cross-environment chroot
is tricky. For example, chroot from glibc to musl results in the errors
below if the iamroot library uses the pthread library:

	$ bash iamroot-shell -c "chroot alpine-minirootfs /lib/ld-musl-x86_64.so.1 bin/busybox"
	Error relocating /home/gportay/src/iamroot/libiamroot.so: __pthread_register_cancel: symbol not found
	Error relocating /home/gportay/src/iamroot/libiamroot.so: __pthread_unregister_cancel: symbol not found

The missing symbols are specific to the glibc library; stubs must be
implemented to satisfy the dynamic loader.

As of today, the iamroot library is built on a glibc world for the glibc
and the musl libraries. Some limited cross-environment tests exist and
remain fragile. Hence, the user has to know exactly how invoking the
dynamic loader for now. Some effort has to be done to run the dynamic
loader straightforward.

However, the support for multiple build was introduced later by the
commit 46fa98db3b79b48508d5fb6fe095cd1fef21f98b. It makes possible to
build many versions of the iamroot library within the exact same source
directory.

This fixes the alpine test that invokes the dynamic loader directly by
building a musl copy of the iamroot library and preloading it using the
command-line argument --preload.