Fix calling missing stat symbols prior to glibc-2.33, once again
The commit 5518acad56ea63e0abab04cff55b1ab8c175bc3f fixes the build
issue with glibc prior to glibc-2.33.

However, that change does not handle the runtime issue, i.e. chrooting
in a glibc-2.33 world (as Fedora 33).

This calls the xstat symbol if the original stat symbol is not loaded
(aka does not exist).

Fixes:

	$ bash iamroot-shell -c "dnf --releasever 33 --assumeyes --installroot /home/gportay/src/iamroot-multi-arch/fedora-33-rootfs group install minimal-environment"
	(...)
	next_lstat: /home/gportay/src/iamroot/libiamroot.so: undefined symbol: lstat
	(...)
	next_fstatat: /home/gportay/src/iamroot/libiamroot.so: undefined symbol: fstatat
	(...)
	next_fstat: /home/gportay/src/iamroot/libiamroot.so: undefined symbol: fstat
	(...)