*stat*: drop preprocessor directive for hidden symbols
The commit s5518acad56ea63e0abab04cff55b1ab8c175bc3f and the later
6bc6c26c42c4eb1aa3ab8ef267c917258796d078 have both fixed the subsequent
calls to some stat symbols that were hidden before the version 2.33 of
glibc.

The library is acting differently if it is built from a glibc 2.33's
host (and previous versions) or from a glibc 2.34's host (and future
versions).

This drops the preprocessor directives to compile a library that have
the same behaviour what ever the glibc's version installed on the host.
The library has to be usable and has to behave identically whatever the
host on which it has been compiled.