execve,exec.sh: run exec.sh using host environment
The commit 77c1e912fb6a589b50899f3cece38a14165a8461 (and its fixup
commit 5a349d9518e8d5dc2f311a12585335c2304349d4) was an attempt to run
the exec.sh withing the chroot'ed environment.

The purpose was to deal with different versions of the glibc; the
usecase as the creation of a glibc-2.34's rootfs (fedora 35) from a
glibc-2.33's host (arch 12/2021).

That commit addressed the error below while building the glibc-2.34's
rootfs on a glibc-2.33's host:

	  Running scriptlet: glibc-2.34-7.fc35.x86_64           28/303
	/usr/sbin/ldconfig: symbol lookup error: /home/gportay/src/iamroot/fedora-35-rootfs/usr/lib64/libc.so.6: undefined symbol: __nptl_change_stack_perm, version GLIBC_PRIVATE
	(...)
	  Running scriptlet: rootfiles-8.1-30.fc35.noarch      303/303
	/sbin/ldconfig: symbol lookup error: /home/gportay/src/iamroot/fedora-35-rootfs/usr/lib64/libc.so.6: undefined symbol: __nptl_change_stack_perm, version GLIBC_PRIVATE
	warning: %transfiletriggerin(glibc-common-2.34-7.fc35.x86_64) scriptlet failed, exit status 127

	Error in <unknown> scriptlet in rpm package rootfiles

It fixed the second error, and it raised the following error as the bash
interpreter is not installed yet in the chroot'ed rootfs at the moment
of installing the glibc package:

	  Running scriptlet: glibc-2.34-7.fc35.x86_64           28/303

	getinterp: No such file or directory
	error: lua script failed: [string "%post(glibc-2.34-7.fc35.x86_64)"]:14: assertion failed!
	error: No exec() called after fork() in lua scriptlet

However, that usecase showed of that it was not the adequate solution to
address that symbol issue as the interpreter used to run the exec.sh may
not exist in the chroot'ed rootfs.

Furthermore, that exec script can be overloaded with IAMROOT_EXEC and it
may be written in a less common language that the shell; the interpreter
may not be installed in a minimal chroot'ed rootfs; bash in a busybox
based world (alpine); and even worst with python and its different
versions.

Therefore, the right approach is to run the exec script and its from the
host environment, that will also simplify the underlaying commands as
they are designed to be run from host for most of them (at least for
now).

The reverts commit 77c1e912fb6a589b50899f3cece38a14165a8461, and renames
the dynamic loader enviroment variables (LD_PRELOAD and LD_LIBRARY_PATH)
in lower case to avoid the host environment to get mixed with libraries
of the chroote'ed world.

Fixes:

	  Running scriptlet: glibc-2.34-7.fc35.x86_64           28/303

	getinterp: No such file or directory
	error: lua script failed: [string "%post(glibc-2.34-7.fc35.x86_64)"]:14: assertion failed!
	error: No exec() called after fork() in lua scriptlet

And addresses the original errors:

	  Running scriptlet: glibc-2.34-7.fc35.x86_64           28/303
	/usr/sbin/ldconfig: symbol lookup error: /home/gportay/src/iamroot/fedora-35-rootfs/usr/lib64/libc.so.6: undefined symbol: __nptl_change_stack_perm, version GLIBC_PRIVATE
	(...)
	  Running scriptlet: rootfiles-8.1-30.fc35.noarch      303/303
	/sbin/ldconfig: symbol lookup error: /home/gportay/src/iamroot/fedora-35-rootfs/usr/lib64/libc.so.6: undefined symbol: __nptl_change_stack_perm, version GLIBC_PRIVATE
	warning: %transfiletriggerin(glibc-common-2.34-7.fc35.x86_64) scriptlet failed, exit status 127

	Error in <unknown> scriptlet in rpm package rootfiles

[1]: 5a349d9518e8d5dc2f311a12585335c2304349d4