Fix setting spurious mode execution bit in file mode
The macros __{,f}warn_if_insuffisant_user_mode{,at} check if the path is
a directory via the functions {,f}isdirectory{,at}() in order to add the
executable bit which may be needed to apply file-system modifications.

The macros assume the function always succeed, and they wrongly consider
the path is a directory even if the function has failed; which happens
if the path does not exist.

This fixes the setting for spurious mode execution bit in file mode by
checking the return value for the functions {,f}isdirectory{,at}()
against a positive value greater than 0; which handles the error case.

This fixes commit e7e0ab2bbb01e96697dd41442524b9f956b34834.