execve: fix check for glibc's dl version
The dynamic loader in glibc worlds may or may not be a symlink depending
on its version or the distribution.

Furthermore, it may have to follow different level of symlinks to reach
to final ld.so binary.

Therfore, it would be better to trust the final filename instead of
checking fo the symlink string.

This resolves the path to the dynamic loader to check for its version
instead of reading the symlink.

Fixes:

	$ iamroot-shell -c "debootstrap impish ubuntu-impish-rootfs http://archive.ubuntu.com/ubuntu"
	(...)
	W: Failure trying to run: chroot "/home/gportay/src/iamroot/ubuntu-impish-rootfs" /bin/true
	W: See /home/gportay/src/iamroot/ubuntu-impish-rootfs/debootstrap/debootstrap.log for details

	$ cat /home/gportay/src/iamroot/ubuntu-impish-rootfs/debootstrap/debootstrap.log
	(...)
	chroot: failed to run command '/bin/true': Operation not supported