path_resolution: no not ignore /run/systemd in ignore()
The regex does not support for negative expressions. The perl regex
(PCRE) has the nevative lookahead or lookbehind syntax that would help
to specify the path to ignore and the path to not ignore (ie. allow) in
the same expression.

The pseudo file-systems mountpoints are ignored (/proc, /sys, /dev and
/run) to avoid mounting them due to a lack of privileges; therefore, the
chroot enviroment reuses the file-systems mounted on host.

Some package scripts running in chroot consider systemd is running by
testing the existence of the directory /run/systemd; they run systemd
commands. But these commands may fail for multiple reasons (for example,
if systemd is not yet installed).

Therefore, the chroot environment would need to ignore the files and the
directories under a mount-point (/run/*) but not necessarly all of them
(/run/systemd). But the regex has no syntax for this.

	^/(...|/run/(!systemd)|...)/

It is not decided if the environment IAMROOT_PATH_RESOLUTION_IGNORE must
list all the ignored sub-paths or if a new environment is to be created
to white list path that are ignored.

This forces not ignoring /run/systemd by hardcoding it in the function
ignore().