fchownat: fix loading ownership of resolved path
The macro __getuid() and __getpid() take the path as unique parameter
but none is given as argument in the function fchownat().

The code generated by the preprocessor use the original path given as
argument to fchownat() instead of the path resolved by path_resolution()
since the name of the path parameter of the macro exists in the calling
scope.

Therefore, the compiler does not fail even if no argument is given to
the macro, and it results in loading the ownership of the path that is
not resolved in the chroot directory.

This fixes commit b4179e6fdabcf654696f8a030e678919355ecc2a.