path_resolution: do not return EINVAL if dfd is -1
The internal function path_resolution() returns EINVAL if the given
directory fd is -1 since the function fpath_resolutionat() was merged to
the function in the commit 8b52cdf775c7ab2cceb764def9d6ff2731fc3fe2..

According to man open(2), openat returns EINVAL:

	EINVAL The filesystem does not support the O_DIRECT flag.  See
	       NOTES for more information.

	EINVAL Invalid value in flags.

	EINVAL O_TMPFILE was specified in flags, but neither O_WRONLY
               nor O_RDWR was specified.

	EINVAL O_CREAT was specified in flags and the final component
	       ("basename") of the new file's pathname is invalid (e.g.,
	       it contains characters not  permitted  by the underlying
               filesystem).

	EINVAL The final component ("basename") of pathname is invalid
	       (e.g., it contains characters not permitted by the
               underlying filesystem).

It does not return EINVAL if fd is -1.

This stops returning EINVAL if the given directory fd is -1 as it is
permitted by the POSIX standard.