path_resolution: preserve errno in __procfdreadlink()
The function realpath() calls readlink() underneath in both glibc[1] and
musl[2] implementations of the libc.

The call to readlink() is subject to change the value of errno, and the
implementation do no reset errno to 0 if it return success.

The musl[2] implementation of realpath() was stolen to implement it
internally. The commit 3b0f492da4ad00983d9c6213784c2ff1231c14f0 saves
and restores the original errno value; see its message for explanations.

The internal function __procfdreadlink() calls the real implementation
of the function readlinkat() that may alter the original errno value.

The internal function path_resolution() has a path that calls indirectly
the function __procfdreadlink(); therefore, it may alter the original
errno value.

Since almost all the intercepted functions call the internal function
path_resolution(), almost all the intercepted functions may alter the
original errno value.

The users of the libc are free to check directly the value of errno, and
bypassing the return code even if this is discouraged. As a consequence,
the library should not alter the errno value. Instead, it should saves
and restores it to keep it untouched.

This saves errno after enterring the function __procfdreadlink() to
restore it after the function readlinkat() returned.

[1]: https://sourceware.org/git/?p=glibc.git;a=blob;f=stdlib/canonicalize.c;hb=glibc-2.37#l188
[2]: https://git.musl-libc.org/cgit/musl/commit/src/misc/realpath.c?id=29ff7599a448232f2527841c2362643d246cee36