path_resolution: resolve /proc/1/root to /
The files /proc/1/{cwd,root,exe} are readable by root only.

The file /proc/1/root is often use to check if the process is in chroot.
systemd stats it in its function running_in_chroot().

However, systemd does not use its internal function in every binaries.
For example, systemctl fails:

	root@archlinux:/# systemctl status
	System has not been booted with systemd as init system (PID 1). Can't operate.
	Failed to connect to bus: Host is down

While it is assumed to success:

	root@archlinux:/# systemctl status
	Running in chroot, ignoring request: status

This assumes the /proc/1/root is a symbolic link to / and resolves it
manually.

Fixes:

	root@archlinux:/# stat /proc/1/root
	  File: /proc/1/root/bin/bash: cannot read symbolic link '/proc/1/root': Permission denied

	  Size: 0         	Blocks: 0          IO Block: 1024   symbolic link
	Device: 13h/19d	Inode: 2429        Links: 1
	Access: (0777/lrwxrwxrwx)  Uid: (    0/    root)   Gid: (    0/    root)
	Access: 2023-04-14 07:35:50.903331217 +0200
	Modify: 2023-04-13 06:37:34.723333328 +0200
	Change: 2023-04-13 06:37:34.723333328 +0200
	 Birth: -