path_resolution: ignores path after symlink expansion
The file /etc/fstab is a relative symlink to ../proc/self/mounts.

The function path_resolution() resolves it $ROOT/proc/self/mounts once
the symlinks are followed.

This expended path should be ignored and resolved to /proc/self/mounts
then if the /proc are ignored, but it is not. The path are ignored
before the symlinks are expanded.

This ignores the resolved path after the symlinks are followed and
expanded.