readlinkat: fix in-chroot truncated symlinks
According to readlinkat(2):

	NOTES

	Using a statically sized buffer might not provide enough room
	for the symbolic link contents. The required size for the buffer
	can be obtained from the stat.st_size value returned by a call
	to lstat(2) on the link. However, the number of bytes written by
	readlink() and readlinkat() should be checked to make sure that
	the size of the symbolic link did not increase between the
	calls. Dynamically allocating the buffer for readlink() and
	readlinkat() also addresses a common portability problem when
	using PATH_MAX for the buffer size, as this constant is not
	guaranteed to be defined per POSIX if the system does not have
	such limit.

The current implementation reads the symlink using the given buffer. The
buffer size can be obtained from lstat(3) as stated by the note above;
i.e. the symlink size with the root directory already stripped.

Therefore, this buffer may be too small to read the whole symlink with
the root directory included. This results in having a truncated symlink
in the given buffer.

This fixes the in-chroot truncated symlink by reading the symlink to a
temporary allocated buffer before copying it back to the given buffer.