Fix "group, passwd, shadow: set errno to ENOENT"
TL;DL; This fixes returning the unexpected ENOENT on last entry by the
functions getpwnam_r(), getpwuid_r(), getpwnam(), getpwuid(),
getpwent(), getgrnam_r(), getgrgid_r(), getgrnam(), getgrgid(),
getgrent() and getgrouplist(). The functions fgetpwent() and fgetgrent()
keep returning ENOENT to satisfy systemd; the shadow functions are left
untouched.

This fixes commit 9b023f6817452d7648ab2c3d0467a01629435697.

systemd-sysusers requires the commit above to creates the users and the
groups successfully[1][2].

systemd implements sane helpers on top of the functions fgetpwent()[3]
and fgetgrent()[4].

These helpers returns EIO if the functions returns a NULL pointer and if
the errno is not set to ENOENT[5][6].

	if (!p && errno != ENOENT)
		return errno_or_else(EIO);

These functions were stolen from the musl library and they do not return
ENOENT. They return NULL and the errno is not set if there are no more
entries.

The man page for fgetpwent(3) tells:

	RETURN VALUE

	The fgetpwent() function returns a pointer to a passwd
	structure, or NULL if there are no more entries or an error
	occurs. In the event of an error, errno is set to indicate the
	error.

	ERRORS

	ENOMEM Insufficient memory to allocate passwd structure.

There is no reference to ENOENT.

The ENOENT errno value may significate either the files /etc/passwd or
/etc/group are not found, or there is no entry (or no more).

The former is false as the stream is already opened by fopen()[7][8].

The library glibc implements the function on top of NSS.

The functions fgetpwent()[9] and fgetgrent()[10] call their reentrant
versions fgetpwent_r()[11] and fgetgrent_r()[12].

Both functions end to the internal __nss_fgetent_r()[13]:

	/* Generic implementation of fget*ent_r.  Reads lines from FP until
	   EOF or a successful parse into *RESULT using PARSER.  Returns 0 on
	   success, ENOENT on EOF, ERANGE on too-small buffer.  */
	int __nss_fgetent_r (FILE *fp, void *result,
	                     char *buffer, size_t buffer_length,
	                     nss_files_parse_line parser) attribute_hidden;

The note tells the function returns ENONENT on EOF, i.e. no more entry
to read. See the function __nss_readline()[14]:

	          if (__feof_unlocked (fp))
	            {
	              __set_errno (ENOENT);
	              return ENOENT;
	            }

The commit 9b023f6817452d7648ab2c3d0467a01629435697 tried to fix this
behaviour by appling the same change to the internal functions stolen
from the musl library.

	rv = ferror(f) ? errno : ENOENT;

Unfortunatly, this behaviour has side effects and breaks the upcoming
function initgroups()[15] that is used by su[16].

	$ su iamroot
	su: cannot set groups: No such file or directory

The musl implementation of the function initgroups() calls indirectly
the internal function __getgrent_a() via the function getgrouplist().
It ends with NULL and errno ENOENT on last entry.

Therefore, the functions getgrouplist() and initgroups() unexpectedly
fail with the error:

	No such file or directory.

This moves the workaround to return NULL and ENOENT on last entry to the
functions fgetpwent() and fgetgrent() only; that satisfy systemd.

Additionally, this applies the same change to fgetspent() that does not
have the internal function __getspent_a().

[1]: https://github.com/systemd/systemd/blob/v253/src/sysusers/sysusers.c#L2250-L2252
[2]: https://github.com/systemd/systemd/blob/v253/src/sysusers/sysusers.c#L2254-L2256
[3]: https://github.com/systemd/systemd/blob/v253/src/sysusers/sysusers.c#L189-L210
[4]: https://github.com/systemd/systemd/blob/v253/src/sysusers/sysusers.c#L237-L258
[5]: https://github.com/systemd/systemd/blob/v253/src/basic/user-util.c#L984-L995
[6]: https://github.com/systemd/systemd/blob/v253/src/basic/user-util.c#L1010-L1021
[7]: https://github.com/systemd/systemd/blob/v253/src/sysusers/sysusers.c#L173
[8]: https://github.com/systemd/systemd/blob/v253/src/sysusers/sysusers.c#L221
[9]: https://sourceware.org/git/?p=glibc.git;a=blob;f=pwd/fgetpwent.c;hb=glibc-2.37#l32
[10]: https://sourceware.org/git/?p=glibc.git;a=blob;f=grp/fgetgrent.c;hb=glibc-2.37#l32
[11]: https://sourceware.org/git/?p=glibc.git;a=blob;f=pwd/fgetpwent_r.c;hb=glibc-2.37#l72
[12]: https://sourceware.org/git/?p=glibc.git;a=blob;f=grp/fgetgrent_r.c;hb=glibc-2.37#l58
[13]: https://sourceware.org/git/?p=glibc.git;a=blob;f=include/nss_files.h;hb=glibc-2.37#l147
[14]: https://sourceware.org/git/?p=glibc.git;a=blob;f=nss/nss_readline.c;hb=glibc-2.37#l48
[15]: https://git.musl-libc.org/cgit/musl/tree/src/misc/initgroups.c?h=v1.2.3
[16]: https://github.com/util-linux/util-linux/blob/v2.38.1/login-utils/su-common.c#L795