catopen: do not resolve none-absolute path as relative
According to catopen(3):

	DESCRIPTION

	The argument name specifies the name of the message catalog to
	be opened. If name specifies an absolute path (i.e., contains a
	'/'), then name specifies a pathname for the message catalog.
	Otherwise, the environment variable NLSPATH is used with name
	substituted for %N (see locale(7)). It is unspecified whether
	NLSPATH will be used when the process has root privileges. If
	NLSPATH does not exist in the environment, or if a message
	catalog cannot be opened in any of the paths specified by it,
	then an implementation defined path is used. This latter default
	path may depend on the LC_MESSAGES locale setting when the flag
	argument is NL_CAT_LOCALE and on the LANG environment variable
	when the flag argument is 0. Changing the LC_MESSAGES part of
	the locale may invalidate open catalog de‐ scriptors.

A none-absolute path is not resolved as a relative to the current
working directory. It is resolved using NFSPATH instead.

That environment variable contains special wildcards prefixed by %, and
it requires some works to expand the placehodlers properly.

Altering the NFSPATH variable by prepending the root directory is way
too invasive and may have unexpected side effects.

This delegates the resolution of none-absolute paths to the next symbol
for now. As a consequence, it resolves the catalog files from the host
instead the one withing the root directory.