execvp, execvP: merge files together
iamroot comes with the two files execvp.c and execvP.c that cannot be
differentiated in a case insensitive filesystem.

The GitLab Actions FreeBSD-vm[1] and OpenBSD-vm[2] uses a MacOS image
and virtualization to run these Operating Systems.

However, the MacOS installation defaults to case insensitive APFS
filesystem, and there is no plan to support a case sensitive filesystem
image at this time "since it would affect overall capacity of Mac and be
costly to maintain"[3].

Therefore, these actions ends with the following error as both files
execvp.c and execvP.c cannot exists in a case insensitive filesystem.

This merges both files together to have a single file and fix the error
below.

Fixes:

	gmake: *** No rule to make target 'execvP.o', needed by 'libiamroot.so'.  Stop.

[1]: https://github.com/vmactions/freebsd-vm
[2]: https://github.com/vmactions/openbsd-vm
[1]: https://github.com/actions/runner-images/issues/865