Remove tmpnam() and tmpnam_r()
This reverts partially commit 048a0323e4ba8f939b119209a96e3059b9f33f4c.

According to man tmpnam(3):

	The tmpnam() function returns a pointer to a string that is a
	valid filename, and such that a file with this name did not
	exist at some point in time, so that naive programmers may think
	it a suitable name for a temporary file. If the argument s is
	NULL, this name is generated in an internal static buffer and
	may be overwritten by the next call to tmpnam(). If s is not
	NULL, the name is copied to the character array (of length at
	least L_tmpnam) pointed to by s and the value s is returned in
	case of success.

	The created pathname has a directory prefix P_tmpdir. (Both
	L_tmpnam and P_tmpdir are defined in <stdio.h>, just like the
	TMP_MAX mentioned below.)

	The tmpnam_r() function performs the same task as tmpnam(), but
	returns NULL (to indicate an error) if s is NULL.

This removes the functions tmpnam() and tmpnam_r() as the parameter is a
buffer of length L_tmpnam that holds the output pathname, i.e. it is not
a pathname to be chroot resolved.

Furthermore, the lenght L_tmpnam is 20 bytes long only; it is not long
enough to hold a chroot resolved pathname.