Add IAMROOT_PATH_RESOLUTION_WARNING_IGNORE
The environment variable IAMROOT_PATH_RESOLUTION_WORKAROUND is a boolean
that does not offer the flexibility to filter the warnings of the path
that contains the root directory already.

Some warnings are due to a bug in the application (as in dnf), some
other warnings are due to a bug in the library, or due to some system
calls that are not intercepted yet.

The new environment variable IAMROOT_PATH_RESOLUTION_WARNING_IGNORE
disables the prepend of the root directory if the path to resolve
contains the root directory already (as for the old environment variable
IAMROOT_PATH_RESOLUTION_WORKAROUND). That new variable allows to filter
the path that contains the root directory already thanks to a regular
expression (as for the two other existing environment variables
IAMROOT_PATH_RESOLUTION_IGNORE and IAMROOT_PATH_RESOLUTION_ALLOW).

This offers the possibility to leave the warning on until the library is
fully operational.

For a reason that is still unknown yet, the syscall readlinkat() is not
intercepted by the library while creating a Fedora 38 on Arch Linux.

The scriptlet systemd runs systemd-sysusers[1] to create groups, and it
throws the warning below if IAMROOT_PATH_RESOLUTION_WARNING_IGNORE is
not set to /var/log/dnf.rpm.log|/etc:

	Warning: $ROOT/x86_64-fedora-38-rootfs/etc: contains root directory '$ROOT/x86_64-fedora-38-rootfs'

The binary systemd-sysusers opens the parent directory of the file
/etc/.#groupXXXXXXXXXXXXXXXX that was opened via fopen(). See the stack
trace below:

	run()
	write_files()
	write_temporary_group()
	fflush_sync_and_check()[2]
	fsync_full()[3]
	fsync_directory_of_file()[4]
	open_parent()[5]
	open()[6]

The path of the parent directory is resolved via the internal function
fd_get_path() and the procfs:

	fd_get_path()[7]
	readlink_malloc()[8]
	readlinkat_malloc()[9]
	readlinkat()[10]

The syscall readlinkat() is not resolved and the root directory is not
stripped by the library, causing the warning mentionned above.

[1]: https://src.fedoraproject.org/rpms/systemd/blob/f38/f/systemd.spec#_886
[2]: https://github.com/systemd/systemd-stable/blob/v253.2/src/sysusers/sysusers.c#L770-L772
[3]: https://github.com/systemd/systemd-stable/blob/v253.2/src/basic/fileio.c#L1169-L1171
[4]: https://github.com/systemd/systemd-stable/blob/v253.2/src/basic/sync-util.c#L79-L81
[5]: https://github.com/systemd/systemd-stable/blob/v253.2/src/basic/sync-util.c#L64-L66
[6]: https://github.com/systemd/systemd-stable/blob/v253.2/src/basic/fs-util.c#805
[7]: https://github.com/systemd/systemd-stable/blob/v253.2/src/basic/sync-util.c#L46
[8]: https://github.com/systemd/systemd-stable/blob/v253.2/src/basic/fd-util.c#L536
[9]: https://github.com/systemd/systemd-stable/blob/v253.2/src/basic/fs-util.c#L153
[10]: https://github.com/systemd/systemd-stable/blob/v253.2/src/basic/fs-util.c#L130