path_resolution: ignore if contains root directory
The resolved path does not contain necessarily the root directory.

It exists a path in which the resolved path has not the root directory
prepended: if the path is relative to the current working directory.

Therefore, the resolved path does not start with the root directory.

This ignores the resolved path if it starts with the root directory
only.