linux.mk: reduce target scope of APK_OPTS
The environment variable APK_OPTS is needed to build the rootfs only,
i.e. for the target /bin/busybox.

This reduces the scope of targets of the environment APK_OPTS by
removing the target -chroot from the exporting of the variable.

Note: The target -shell it is used to run shell ready to run the command
to create the rootfs, and it should have the variable set.