dso: rename parameter dt_tag to d_tag
The functions __dl_iterate_ehdr{,32,64}() uses the dt_tag parameter to
compare it to the dynamic entry type d_tag of Elf{32,64}_Dyn (i.e.
DT_NEEDED, DT_RUNPATH, DT_RPATH...); therefore the name dt_tag should be
the same of the one in the structure.

This renames the argument dt_tag in the functions __dl_iterate_ehdr*()
to d_tag.