dlopen, dlmopen: open needed shared objects manually
The symbol dlopen() from the libc resolves and opens the needed shared
objects before the given shared object is actually opened.

It results in loading the shared objects from the host instead of the
ones from the chroot'ed environment.

This opens the needed shared objects manually, before the given shared
object is opened.

Note: It implements the dynamic loader searches as defined in dlopen(3).