dso: guess the deflib from the shared object
According to ld.so(8):

	On some 64-bit architectures, the default paths for 64-bit
	shared objects are /lib64, and then /usr/lib64.

The default library path differs on some 64-bit architectures on the
GNU/Linux systems. This behaviour is driven by the environment variable
IAMROOT_LIBRARY_PATH.

The glibc x86_64 and aarch64 architectures use the directory lib64
instead of lib (i.e. IAMROOT_LIBRARY_PATH=/lib64:/usr/lib64).

The musl library and both FreeBSD and OpenBSD systems use directory lib
only.

Furthermore, this default library path is distro specific on the Linux
systems. Arch Linux (x86_64 only) uses lib, symlinks lib64 to lib and
uses lib32 for its multilib support. Fedora uses distinct directories
for both lib and lib64, lib for 32-bits, lib64 64-bit. It is different
in the Debian world and its multiarch[1] support; it adds a tuple[2]
directory after the lib directory for the architecture.

This makes the magical very tricky to guess the default library path on
the Linux systems; it shall support the following situations:
 1. cross-chroot libc (i.e. from GNU World to musl)
 2. cross-chroot architecture (i.e. form x86-64 to i686 or armv7-a)
 3. execve executables (i.e. shared object with an interpreter)
 4. dlopen libraries (i.e. shared object without an interpreter)

The magic is based on the ELF header to guess if the chroot is a 32-bit
or a 64-bit world and if the operating system and its ABI is a either
UNIX System V or GNU/Linux or even FreeBSD.

The name of the dynamic loader is also needed to detect a Linux world
since the GNU/Linux ELF shared objects can be either UNIX System V or
GNU/Linux (OpenBSD uses UNIX System V as well).

The dynamic loader is in the interpreter segment of the ELF executable
file. However, the none-executable files ELF shared objects (such as
libraries) does not have that segment.

Therefore, it is hard to determine if the chroot world is either a
64-bit GNU/Linux or a musl (or even OpenBSD), and if it has to use
either /lib64:/usr/lib64 or /lib:/usr/lib as default library path
though; as needed by the point 4.

The libc soname is system specific:
 - libc.so.6 for GNU/Linux since glibc 2.0
 - libc.so.5 for Linux libc (former libc based on glibc 1; see the note
   below)
 - libc.so for musl (note: the dynamic loader is a symlink to libc)
 - libc.so.5 for FreeBSD 5.0
 - libc.so.6 for FreeBSD 6.0
 - libc.so.7 for FreeBSD since since 7.0
 - libc.so.96.2 for OpenBSD 7.2
 - libc.so.97.1 for OpenBSD 7.3

It is not ideal to rely on the libc soname as it is subject to collision
between the different operating systems; for example with the libc.so.6
of FreeBSD 6.x.

Hopefully, the libc.so.6 soname is pretty stable as the glibc 2.0 was
released in the early of 1997 (i.e. 26 years ago)[3].

Even better, the GNU libc needs the dynamic loader while the FreeBSD
libc does not; a least since 2.0.7 (tested down to Debian 2.0 Hamm[4];
Debian 1.3 Bo[5] was using the former Linux libc fork, aka libc.so.5).

Debian 2.0 (i386):

	$ curl -O http://archive.debian.org/debian/dists/hamm/main/binary-i386/base/libc6_2.0.7t-1.deb
	$ ar x libc6_2.0.7t-1.deb
	$ tar xf data.tar.gz
	$ readelf -a lib/libc.so.6 | grep -E '(NEEDED|SONAME)'
	 0x00000001 (NEEDED)                     Shared library: [ld-linux.so.2]
	 0x0000000e (SONAME)                     Library soname: [libc.so.6]
	$ file -L lib/libc.so.6
	lib/libc-2.0.7.so: ELF 32-bit LSB shared object, Intel 80386, version 1 (SYSV), dynamically linked, interpreter /lib/ld-linux.so.2, stripped, too many notes (256)
	$ lib/ld-linux.so.2 lib/libc.so.6
	GNU C Library production release version 2.0.7, by Roland McGrath et al.
	Compiled by GNU CC version 2.7.2.3.
	Copyright (C) 1992, 93, 94, 95, 96, 97, 98 Free Software Foundation, Inc.
	This is free software; see the source for copying conditions.
	There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A
	PARTICULAR PURPOSE.
	Compiled on a Linux 2.0.33 system on 1998/07/16.
	Available extensions:
		GNU libio by Per Bothner
		BIND-4.9.7-REL
		NIS(YP) NSS modules 0.8 by Thorsten Kukuk
		UFC-crypt, patchlevel 1e by Michael Glad
		linuxthreads-0.6 by Xavier Leroy
	Report bugs using the `glibcbug' script to <bugs@gnu.org>.

Debian 1.3 (i386):

	$ curl -O http://archive.debian.org/debian/dists/bo/main/binary-i386/base/libc5_5.4.33-6.deb
	$ ar x libc5_5.4.33-6.deb
	$ tar xf data.tar.gz
	$ readelf -a lib/libc.so.5 | grep -E '(NEEDED|SONAME)'
	 0x0000000e (SONAME)                     Library soname: [libc.so.5]

Consequently, the default library path may be guessed to dlopen the
shared objects that are not executable files but that are linked against
the GNU libc; as long as the libc.so.6 is the library soname and as long
as it is executable and contains the needed dynamic loader. This hacky
guess has to be updated after every bump of the libc soname or if the
libc ceases to be executable (i.e. no more need to the dynamic loader or
no more interpreter).

It falls back to the executable file if the shared object is not linked
against the GNU libc library.

This guesses the default library path of the chroot'ed Linux environment
by doing the magic mentionned above.

Note: According to libc(7):

	Linux libc

	In the early to mid 1990s, there was for a while Linux libc, a
	fork of glibc 1.x created by Linux developers who felt that
	glibc development at the time was not sufficing for the needs of
	Linux. Often, this library was referred to (ambiguously) as just
	“libc”. Linux libc released major versions 2, 3, 4, and 5, as
	well as many minor versions of those releases. Linux libc4 was
	the last version to use the a.out binary format, and the first
	version to provide (primitive) shared library support. Linux
	libc 5 was the first version to support the ELF binary format;
	this version used the shared library soname libc.so.5. For a
	while, Linux libc was the standard C library in many Linux
	distributions.

	However, notwithstanding the original motivations of the Linux
	libc effort, by the time glibc 2.0 was released (in 1997), it
	was clearly superior to Linux libc, and all major Linux
	distributions that had been using Linux libc soon switched back
	to glibc. To avoid any confusion with Linux libc versions, glibc
	2.0 and later used the shared library soname libc.so.6.

	Since the switch from Linux libc to glibc 2.0 occurred long ago,
	man-pages no longer takes care to document Linux libc details.
	Nevertheless, the history is vis‐ ible in vestiges of
	information about Linux libc that remain in a few manual pages,
	in particular, references to libc4 and libc5.

[1]: https://wiki.debian.org/Multiarch
[2]: https://wiki.debian.org/Multiarch/Tuples
[3]: https://linux.die.net/man/7/libc
[4]: http://archive.debian.org/debian/dists/hamm/main/binary-i386/base/libc6_2.0.7t-1.deb
[5]: http://archive.debian.org/debian/dists/bo/main/binary-i386/base/libc5_5.4.33-6.deb